Fix detection of IPv6 IETF Protocol Assignments address block
When ipcalc was built with the "-Wall" flag with recent versions of GCC
(at least with GCC 9.2.1 on Debian, but probably with a few previous
releases of GCC too), which has been the default since the inception of
the Meson build and since commit e7a6cb2 in the Makefile, it complained
with the following warning message:

[1/3] Compiling C object 'ipcalc@exe/ipcalc.c.o'.
../ipcalc.c: In function ‘ipv6_net_to_type’:
../ipcalc.c:799:23: warning: bitwise comparison always evaluates to false [-Wtautological-compare]
  799 |  if ((word1 & 0xfffe) == 0x2001 && word2 == 0)
      |                       ^~

Although the CI builds defined in .gitlab-ci.yml all build ipcalc with
the "-Werror" flag on a recent version of Fedora, which theoretically
should have turned the "-Wtautological-compare" warning into an error
and failed the build, they omitted the "-Wall" flag, so GCC didn't emit
that particular warning.

It turns out that GCC was correct, and the check for the IPv6 IETF
Protocol Assignments address space in ipcalc was wrong according to the
definition in RFC2928 (https://tools.ietf.org/html/rfc2928). This commit
fixes the bug by comparing the whole first word of the IPv6 address and
the lower nine bits of the second word, along with its prefix, like many
of the other address block tests do. I also added a new unit test for
the IETF Protocol Assignments address space to prove that it is being
calculated correctly, both now and to prevent regressions in the future.

Signed-off-by: Karl Lenz <xorangekiller@gmail.com>