Fix detection of IPv6 6to4 address block
While fixing the IPv6 IETF Protocol Assignments address block detection
in the last commit, I noticed that the IPv6 6to4 address block detection
was also broken while reviewing the code. Unlike with the IETF Protocol
Assignments address block, there was no warning message emitted by the
compiler, but because of where the check was placed in relation to the
much more broad Global Unicast address block detection, it was
impossible to ever declare an IPv6 address to be in the 6to4 address
block even if it really was. It would always be declared to be in the
Global Unicast address block instead. This bug has been present ever
since the 6to4 detection was added in commit 24d63d8.

I didn't have to change the IPv6 6to4 address block detection logic in
this commit to fix the bug. I merely had to move it above the Global
Unicast address block detection. I also added a new unit test to prove
that it works correctly now, and to hopefully prevent any regressions of
this sort in the future.

Signed-off-by: Karl Lenz <xorangekiller@gmail.com>