Fix TEST-NET-2 detection
I noticed that detection of the IPv4 special-purpose TEST-NET-1 address
block was fixed recently in commit 038f45e, but it didn't have any unit
tests. That should be a fairly simple thing to do, so I added a new
unit test for it as well as the logically related (but not adjacent)
TEST-NET-2 and TEST-NET-2 according to the latest version of the
IANA IPv4 Special-Purpose Address Registry, and I found another
problem. Similar to how bytes 2 and 3 were swapped in TEST-NET-1 until
recently, the first byte of TEST-NET-2 was wrong. It was "192" like
TEST-NET-1 rather than "198" as dictated by RFC5737. This commit fixes
that problem and adds unit test cases for all three address blocks.

See: <https://tools.ietf.org/html/rfc5737>

Signed-off-by: Karl Lenz <xorangekiller@gmail.com>