Fix ULA prefix generator to use only defined ULA range
According to RFC 4193 the format of ULA prefix is fc00::/7 followed by L
bit which has to be set to 1 for locally assigned prefixes. The meaning
of L flag reset to zero may be defined in the future, but is now
undefined. Therefore, only defined ULA addresses are in fd00::/8 range.

Fixes #27.

Signed-off-by: Ondřej Caletka <ondrej@caletka.cz>